🔧 refactor: inline image model params code to models file
