+++ src/config/aiModels/openai.ts
@@ -1,4 +1,4 @@
-import { gptImage1ParamsSchema } from '@/config/paramsSchemas/openai/gpt-image-1';
+import { ModelParamsSchema } from '@/libs/standard-parameters';
 import {
   AIChatModelCard,
   AIEmbeddingModelCard,
@@ -7,6 +7,15 @@ import {
   AISTTModelCard,
   AITTSModelCard,
 } from '@/types/aiModel';
+
+export const gptImage1ParamsSchema: ModelParamsSchema = {
+  imageUrls: { default: [] },
+  prompt: { default: '' },
+  size: {
+    default: 'auto',
+    enum: ['auto', '1024x1024', '1536x1024', '1024x1536'],
+  },
+};
 
 export const openaiChatModels: AIChatModelCard[] = [
   {
