--- src/config/paramsSchemas/openai/gpt-image-1.ts
@@ -1,10 +1,0 @@
-import { type ModelParamsSchema } from '@/libs/standard-parameters/index';
-
-export const gptImage1ParamsSchema: ModelParamsSchema = {
-  imageUrls: { default: [] },
-  prompt: { default: '' },
-  size: {
-    default: 'auto',
-    enum: ['auto', '1024x1024', '1536x1024', '1024x1536'],
-  },
-};
