+++ src/store/image/slices/generationConfig/action.test.ts
@@ -1,7 +1,7 @@
 import { act, renderHook } from '@testing-library/react';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { fluxSchnellParamsSchema } from '@/config/paramsSchemas/fal/flux-schnell';
+import { fluxSchnellParamsSchema } from '@/config/aiModels/fal';
 import {
   ModelParamsSchema,
   RuntimeImageGenParams,
